requirements.txt
byol-pytorch
pytorch-lightning
tqdm
wandb
